agent/nano/requirements.nano.txt
# FastAPI Requirements for Jetson Nano (with GPU/CUDA support)
# Optimized for ARM64 architecture and comprehensive health checking

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
starlette==0.27.0

# HTTP and networking
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.0

# JSON and data handling
orjson==3.9.10

# System monitoring (lightweight)
psutil==5.9.6

# Database connectivity
psycopg2-binary==2.9.7

# Environment and configuration
python-dotenv==1.0.0

# Scientific computing (for ML capabilities)
numpy==1.24.4
scipy==1.10.1
pandas==2.0.3
scikit-learn==1.2.2

# Logging and utilities
structlog==23.2.0
click==8.1.7

# Development and debugging
python-multipart==0.0.6

# GPU monitoring and CUDA support
# Note: PyTorch and TensorFlow come from base L4T container
# Note: TensorRT comes from system installation